refactor: rename preset-inspectors -> builtin-inspectors
