+++ src/builtin-inspectors/as-assertions.ts
